Move skeleton to loading.tsx
